requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-filter==25.1
django-cors-headers==4.7.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
